refactored static component
